% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{RECA-package}
\alias{RECA-package}
\title{Relevant Component Analysis}
\description{
The RECA package performs Relevant Component Analysis (RCA).
}
\details{
\tabular{ll}{ Package: \tab RECA\cr Type: \tab Package\cr
Version: \tab 1.0\cr License: \tab GPL (>= 2)\cr }
}
\note{
Bug reports and feature requests could be sent to
\url{https://github.com/road2stat/RECA/issues}.
}
\author{
Nan Xiao <\email{road2stat@gmail.com}>
}

