% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA-package.R
\docType{package}
\name{RECA-package}
\alias{RECA}
\alias{RECA-package}
\title{RECA: Relevant Component Analysis for Supervised Distance Metric Learning}
\description{
Relevant Component Analysis (RCA) tries to find a linear
    transformation of the feature space such that the effect of irrelevant
    variability is reduced in the transformed space.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nanx.me/RECA/}
  \item \url{https://github.com/nanxstats/RECA}
  \item Report bugs at \url{https://github.com/nanxstats/RECA/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{me@nanx.me} (0000-0002-0250-5673)

}
\keyword{internal}
