% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA-package.R
\docType{package}
\name{RECA-package}
\alias{RECA-package}
\title{Relevant Component Analysis}
\description{
The RECA package offers an R implementation for
Relevant Component Analysis (RCA).
}
\details{
\tabular{ll}{ Package: \tab RECA\cr Type: \tab Package\cr
Version: \tab 1.3\cr License: \tab GPL (>= 2)\cr }
}
\note{
Bug reports and feature requests could be sent to
\url{https://github.com/road2stat/RECA/issues}.
}

