% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA-package.R
\docType{package}
\name{RECA-package}
\alias{RECA-package}
\title{Relevant Component Analysis}
\description{
The RECA package offers an R implementation for
Relevant Component Analysis (RCA).
}
\details{
\tabular{ll}{ Package: \tab RECA\cr
Type: \tab Package\cr
License: \tab GPL-3\cr }
}
\note{
Bug reports and feature requests can be sent to
\url{https://github.com/road2stat/RECA/issues}.
}
